[material-ui][Autocomplete] Unselect options when mouse leaves.
What is the problem:
- Currently, when your mouse leaves the Autocomplete Popover, the last
hovered item is selected.

What are we fixing:
- We are adding a minor onMouseLeave event to each option. When you
leave the option, it will remove any selection by selecting the -1
index.

- We also added a prop: keepSelectedOnLeave if people still want this
functionality. But we default to not having this functionality.

I'm open to preserving backward compatibility, but it seems to make
sense from reading the GitHub issues that it would be preferred as not
default behavior.

Note: It might be better performance wise, to add a toplevel mousemove
and mouseleave event instead of a mousemove/mouseleave event per
option. That way we wouldn't have to install so many event listeners.