[core] Rename OpenCollective
Execute on https://github.com/mui/material-ui/pull/39640#discussion_r1382489528. The first step is to make it clear that this is an umbrella OpenCollective, and not the OpenCollective for Material UI. We will use to dedicated open-source collectives once we find it clear that something beyond Material UI works and needs its own support.