[joy-ui][docs] Optimize side navigation's relative contrast
The right sidebar has been updated to the ‘solid’ variant for improved interaction with IconButtons.
It is now darker than the left sidebar, creating better relative contrast and lifting the left
sidebar higher on the screen. Additionally, the styles are consistent whether in light or dark mode,
addressing previous issue where the right sidebar appeared higher/lower based on the color
scheme.