[pigment-css] Setup bundlesize check
Also run the above check in CI.

At the moment, it is only configured to run for @pigment-css/react
but can be extended to be used in any of the other packages.

This is meant to keep the sizes in check over time and fail CI
everytime it increases beyond allowed threshold to allow introspection.