Fix highlighted index doesn't get reset if options length is unchanged (mui#40184)
The bug was that when the box had a limit, after selecting an option and pressing enter if I used the down arrow
key the selected option wouldn't be the first one as expected. The bug came from the "filteredOptions.length"
dependency in "syncHighlightedIndex" because as the list had a limit smaller than the total number of options,
when an option was selected the size of the list didn't change.
For the same reason, the "syncHighlightedIndex" function wasn't called and the index wasn't reset.
To solve the bug, I changed the "filteredOptions.length" dependency to "filteredOptions", so that whenever there
is a change in the "filteredOptions" list, such as removing an element, the function will be called and the index
reset.