Update Autocomplete.js
in this issue https://github.com/mui/material-ui/issues/41780 The startAdornment behavior used in an Autocomplete doesn't match that of a regular TextField. The Autocomplete wraps onto a new line when it gets too small. i think it is a simple fix just to change flexwrap from wrap to nowrap but i can't run the library locally on my computer to test it 

Signed-off-by: fahdmk <51023058+fahdmk@users.noreply.github.com>