Update ResponsiveDrawer.js
In this proposed change I use useMediaQuery and useTheme hooks to make the Drawer component responsive. I am using the Drawer component once instead of how it used twice in docs. I change the variant of the Drawer according to the screen width using the useMediaQuery and useTheme hooks. Just trying to implement it with the DRY principal. One drawer for both mobile and Desktop shown according to Device. If Desktop then the Drawer is permanent and if the Device is a mobile Device then the SAME Drawer is Temporary.

Signed-off-by: MuiaAlex <alexmuia31@gmail.com>