[code-infra] Bump eslint to v9
This still uses the same eslint configuration file by setting
ESLINT_USE_FLAT_CONFIG to `false` to reduce the scope of the PR, plus we
are blocked on [this](https://github.com/airbnb/javascript/issues/2961)
to be implemented.

Other change is the swapping of `eslint-plugin-filenames` to
`eslint-plugin-consistent-default-export-name` since the previous one
was creating issue when running the linter.